+++ include/pytorch/tokenizers/bpe_tokenizer_base.h
@@ -33,7 +33,7 @@ namespace detail {
 using TokenMap = StringIntegerMap<>;
 
 template <typename TToken, typename TRank>
-static Result<TokenMap> buildTokenMap(
+static Result<TokenMap> build_token_map(
     std::vector<std::pair<TToken, TRank>> container) {
   static_assert(
       std::is_same_v<TToken, std::string> ||
@@ -82,7 +82,7 @@ static Result<TokenMap> buildTokenMap(
 };
 
 template <typename TContainer, typename TTokenAccessor, typename TRankAccessor>
-static Result<TokenMap> buildTokenMap(
+static Result<TokenMap> build_token_map(
     const TContainer& container,
     TTokenAccessor token_accessor,
     TRankAccessor rank_accessor) {
@@ -103,7 +103,7 @@ static Result<TokenMap> buildTokenMap(
     pairs.emplace_back(token_accessor(value), rank_accessor(value));
   }
 
-  return buildTokenMap(std::move(pairs));
+  return build_token_map(std::move(pairs));
 }
 
 inline Result<std::unique_ptr<IRegex>> build_special_token_regex(
@@ -123,73 +123,6 @@ inline Result<std::unique_ptr<IRegex>> build_special_token_regex(
     return static_cast<std::unique_ptr<IRegex>>(nullptr);
   }
   return create_regex(special_pattern);
-}
-
-// Hash function for std::pair<uint64_t, uint64_t>
-struct PairHash {
-  std::size_t operator()(const std::pair<uint64_t, uint64_t>& p) const {
-    return std::hash<uint64_t>{}(p.first) ^
-        (std::hash<uint64_t>{}(p.second) << 1);
-  }
-};
-
-// Type alias for BPE merge map: (token_id_1, token_id_2) -> (rank,
-// merged_token_id)
-using MergeMap = std::unordered_map<
-    std::pair<uint64_t, uint64_t>,
-    std::pair<uint64_t, uint64_t>,
-    PairHash>;
-
-// Utility function to build merge ranks map from merge rules
-template <typename TMergeMap>
-inline Result<TokenMap> buildMergeRanksMap(
-    const TMergeMap& merge_map,
-    const TokenMap& token_map) {
-  // Static assertions to verify TMergeMap has the expected key and value types
-  using KeyType = typename TMergeMap::key_type;
-  using ValueType = typename TMergeMap::mapped_type;
-
-  static_assert(
-      std::is_same_v<KeyType, std::pair<uint64_t, uint64_t>>,
-      "TMergeMap key type must be std::pair<uint64_t, uint64_t>");
-
-  static_assert(
-      std::is_same_v<ValueType, std::pair<uint64_t, uint64_t>>,
-      "TMergeMap value type must be std::pair<uint64_t, uint64_t>");
-
-  // Use a map to handle duplicates - keep the lowest rank (highest priority)
-  std::unordered_map<std::string, uint64_t> unique_merge_ranks;
-
-  for (const auto& [pair, rank_and_id] : merge_map) {
-    uint64_t first_id = pair.first;
-    uint64_t second_id = pair.second;
-    uint64_t rank = rank_and_id.first;
-
-    // Get the token strings for the pair
-    auto first_token = token_map.tryGetString(first_id);
-    auto second_token = token_map.tryGetString(second_id);
-
-    if (first_token && second_token) {
-      std::string merged_token =
-          std::string(*first_token) + std::string(*second_token);
-
-      // Keep the entry with the lowest rank (highest priority in BPE)
-      auto it = unique_merge_ranks.find(merged_token);
-      if (it == unique_merge_ranks.end() || rank < it->second) {
-        unique_merge_ranks[merged_token] = rank;
-      }
-    }
-  }
-
-  // Convert to vector for buildTokenMap
-  std::vector<std::pair<std::string, uint64_t>> merge_rank_pairs;
-  merge_rank_pairs.reserve(unique_merge_ranks.size());
-
-  for (const auto& [token, rank] : unique_merge_ranks) {
-    merge_rank_pairs.emplace_back(token, rank);
-  }
-
-  return buildTokenMap(std::move(merge_rank_pairs));
 }
 
 class BPETokenizerBase : public Tokenizer {
@@ -224,6 +157,9 @@ class BPETokenizerBase : public Tokenizer {
       const TokenMap& encoder) const;
 
   // Virtual method for BPE merging - can be overridden by derived classes
+  // The passed in `ranks` param for the base impl is just a regular token map
+  // and that the actual ranks are derived implicitly from the regular token
+  // map. This is the same implementation as Tiktoken.
   virtual std::vector<uint64_t> _byte_pair_merge(
       const std::string& piece,
       const TokenMap& ranks,
@@ -233,7 +169,6 @@ class BPETokenizerBase : public Tokenizer {
   std::unique_ptr<IRegex> special_token_regex_;
   std::optional<TokenMap> token_map_;
   std::optional<TokenMap> special_token_map_;
-  std::unique_ptr<MergeMap> merge_map_;
 
  private:
   virtual Error _encode(
